requirements.txt
streamlit==0.63.0
qrcode==6.1
matplotlib==3.3.0
Pillow==7.2.0
